bug fixed

+++ Akasha/src/org/firo/akasha/CurrentActivity.java
@@ -10,21 +10,20 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
-import android.widget.Button;
 import android.widget.ListView;
 import android.widget.SimpleCursorAdapter;
 import android.widget.AdapterView.OnItemClickListener;
+import android.widget.Toast;
 
 public class CurrentActivity extends Activity {
 
-	public static  CurrentDbHelper currentDb = null;
-	
+	public static CurrentDbHelper currentDb = null;
+
 	public Cursor currentDbCursor;
 	public ListView currentListView;
 	public int _id;
-	
-	
-	
+	Toast toast;
+
 	protected final static int MENU_ADD = Menu.FIRST;
 	protected final static int MENU_EDIT = Menu.FIRST + 1;
 	protected final static int MENU_DELETE = Menu.FIRST + 2;
@@ -63,24 +62,16 @@ public class CurrentActivity extends Activity {
 	public void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.current_layout);
-		if(currentDb == null){
+		if (currentDb == null) {
 			currentDb = new CurrentDbHelper(this);
-			System.out.println("=======================create database =================");
+			System.out
+					.println("=======================create database =================");
 		}
-		currentDb.insert(112, "dlnu", "dlnus");
-		currentDb.insert(111, "suan fad dao lun", "algorithms");
-		currentDb.insert(11, "kernel", "hobby");
+		currentDb.insert(112, "dlnu", "dlnus","112  dlnu \n dlnus");
+		currentDb.insert(111, "suan fad dao lun", "algorithms" ,"111  suan fa dao lun \nalgorithms");
+		currentDb.insert(11, "kernel", "hobby","11 kernel\nhobby");
 		currentListView = (ListView) findViewById(R.id.current_listview);
-//		currentDbCursor = currentDb.select();
-//		@SuppressWarnings("deprecation")
-//		SimpleCursorAdapter currentAdapter = new SimpleCursorAdapter(this,
-//				R.layout.currentdumb, currentDbCursor,
-//				new String[] { CurrentDbHelper.FIELD_ACTION,CurrentDbHelper.FIELD_DESCRIPTION},
-//				new int[] { R.id.topTextView1,R.id.topTextView2 });	
-//		
-//		cu rrentListView.setAdapter(currentAdapter);
-		
-		
+
 		currentListView.setOnItemClickListener(new OnItemClickListener() {
 			public void onItemClick(AdapterView<?> parent, View view,
 					int position, long id) {
@@ -89,23 +80,10 @@ public class CurrentActivity extends Activity {
 				_id = currentDbCursor.getInt(0);
 				if (((ListView) parent).getTag() != null) {
 					((View) ((ListView) parent).getTag())
-							.setBackgroundDrawable(null);
+							.setBackgroundColor(Color.BLACK);
 				}
 				((ListView) parent).setTag(view);
-				view.setBackgroundColor(Color.CYAN);// R.drawable.ic_launcher
-			}
-		});
-
-		Button addTimeTaskButton = (Button) findViewById(R.id.add_button);
-		addTimeTaskButton.setOnClickListener(new View.OnClickListener() {
-
-			public void onClick(View view) {
-				Intent intent;
-				intent = new Intent().setClass(CurrentActivity.this,
-						CurrentAddDialogActivity.class);
-				intent.putExtra("_id", String.valueOf(-1));
-				startActivity(intent);
-				System.out.println("Just for you~~");
+				view.setBackgroundColor(Color.CYAN);
 			}
 		});
 	}
@@ -119,30 +97,48 @@ public class CurrentActivity extends Activity {
 			startActivity(intent);
 		}
 		if (cmd == "edit") {
-			Intent intent;
-			intent = new Intent().setClass(CurrentActivity.this,
-					CurrentModifyDialogActivity.class);
-			System.out.println("id form edit " + _id +"\n");
-			intent.putExtra("_id", _id);
-			startActivity(intent);
+
+			if (currentListView.getTag() == null) {
+				toast = Toast.makeText(getApplicationContext(),
+						"Modify? please select anyone~~", Toast.LENGTH_LONG);
+				toast.show();
+			} else {
+				Intent intent;
+				intent = new Intent().setClass(CurrentActivity.this,
+						CurrentModifyDialogActivity.class);
+				System.out.println("id form edit " + _id + "\n");
+				intent.putExtra("_id", _id);
+				startActivity(intent);
+				currentListView.setTag(null);
+			}
 		}
 
-		if (cmd == "delete")
-			currentDb.delete(_id);
+		if (cmd == "delete") {
+			if (currentListView.getTag() == null) {
+				toast = Toast.makeText(getApplicationContext(),
+						"Delete? please select anyone~~", Toast.LENGTH_LONG);
+				toast.show();
+			} else {
+				((View) currentListView.getTag())
+						.setBackgroundColor(Color.BLACK);
+				currentDb.delete(_id);
+				currentListView.setTag(null);
+			}
+		}
 		currentDbCursor.requery();
 		currentListView.invalidateViews();
 		_id = 0;
 	}
-	
-    public void	onResume(){
-    	super.onResume();
-    	currentDbCursor = currentDb.select();
+
+	public void onResume() {
+		super.onResume();
+		currentDbCursor = currentDb.select();
 		@SuppressWarnings("deprecation")
 		SimpleCursorAdapter currentAdapter = new SimpleCursorAdapter(this,
 				R.layout.currentdumb, currentDbCursor,
-				new String[] { CurrentDbHelper.FIELD_ACTION,CurrentDbHelper.FIELD_DESCRIPTION},
-				new int[] { R.id.topTextView1,R.id.topTextView2 });	
-		
+				new String[] { CurrentDbHelper.FIELD_SHOW },
+				new int[] { R.id.topTextView1 });
+
 		currentListView.setAdapter(currentAdapter);
 	}
 }
